requirements.txt
altair==6.0.0
annotated-types==0.7.0
anthropic==0.88.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
flatbuffers==25.12.19
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h3==4.4.2
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kerykeion==5.12.7
MarkupSafe==3.0.3
narwhals==2.18.1
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.4
plotly==6.6.0
protobuf==7.34.1
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyswisseph==2.10.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.1
requests-cache==1.3.1
rpds-py==0.30.0
scour==0.38.2
simple-ascii-tables==1.0.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.56.0
tenacity==9.1.4
timezonefinder==8.2.2
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.3
watchdog==6.0.0
